requirements.txt
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
plotly>=5.17.0,<7.0.0
openpyxl>=3.1.0,<4.0.0
pytest>=7.4.0,<10.0.0
pytest-cov>=4.1.0,<8.0.0
python-dotenv>=1.0.0,<2.0.0
